fix(lint): remove setState-in-effect violations blocking npm run check
Replace synchronous setState inside effects with the React-recommended
render-phase adjustment pattern (prev-value guard) across:

- ProductFormFields: TextField/TextAreaField draft resets
- ProductTab: per-project mode + loading resets; interview greeting
  refresh keyed on the language prop; DocsUpload initial load uses the
  promise-callback lifecycle pattern
- Settings: sync logic extracted to useSettingsSync hooks
  (useApiEndpointField, useBrandForm) to stay within complexity budget

DocsUpload + DocStatusBadge move to ProductDocsUpload.tsx to keep
ProductTab under the max-lines budget (same convention as the earlier
ProductFormFields extraction). Behavior preserved — all existing
component tests pass unchanged.
